# api/openapi.yaml
openapi: 3.0.3
info:
  title: TaxForward API
  version: 1.0.0
  description: |
    Single source of truth for the contracts between the TaxForward backend
    services (Spring Boot + FastAPI) and the web `client`.

    **Public API** (tags `Documents`, `Invoices`, `Suggestions`, `Chat`) is the
    surface the web client consumes. All public traffic flows through the
    **Traefik** gateway, which validates a Firebase ID token via the
    `auth-service` forward-auth middleware and injects the resolved user id as the
    `X-User-Sub` header before proxying to the owning service. Clients therefore
    only send `Authorization: Bearer <firebase-id-token>`.

    **Internal** service-to-service contracts use gRPC, not REST — their
    protobuf definitions live in `api/proto/` and are not covered by this spec.
servers:
  - url: /
    description: >-
      Public Traefik gateway; the base path is relative to the deployed host
      (e.g. http://localhost in docker-compose dev). The web client calls these
      paths relative to `VITE_API_BASE_URL`.
...
security:
  - firebaseToken: []
paths:
...
  /api/v1/exports/years:
    get:
      tags: [Exports]
      operationId: listExportYears
      summary: Tax years the user has accepted invoices for
      description: >-
        Descending. Invoices without an `invoiceDate` are not attributable to a
        tax year and are omitted.
      responses:
        "200":
          description: List of years
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/exports/summary:
    get:
      tags: [Exports]
      operationId: getExportSummary
      summary: Per-category totals for one tax year
      description: >-
        The figures rendered into the PDF summary, exposed separately so the
        client can preview an export before downloading it.
      parameters:
        - $ref: "#/components/parameters/YearQuery"
      responses:
        "200":
          description: Summary of the year's accepted invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/exports/csv:
    get:
      tags: [Exports]
      operationId: downloadExportCsv
      summary: Download the year's accepted invoices as CSV
      description: >-
        RFC 4180, UTF-8 with a byte-order mark so spreadsheet software reads the
        German umlauts correctly. One row per invoice line item.
      parameters:
        - $ref: "#/components/parameters/YearQuery"
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/exports/pdf:
    get:
      tags: [Exports]
      operationId: downloadExportPdf
      summary: Download the year's Werbungskosten summary as PDF
      description: >-
        A German-language summary sheet: totals per deduction category, the
        itemized invoices behind them, and the amount exceeding that year's
        Arbeitnehmer-Pauschbetrag.
      parameters:
        - $ref: "#/components/parameters/YearQuery"
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/exports/zip:
    get:
      tags: [Exports]
      operationId: downloadExportZip
      summary: Download the complete tax-year package as a ZIP archive
      description: >-
        Contains the PDF summary, the CSV and JSON renderings of the invoices,
        and every original receipt the invoices were extracted from.
      parameters:
        - $ref: "#/components/parameters/YearQuery"
      responses:
        "200":
          description: ZIP archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Internal service-to-service calls (latest-invoices, embed/delete) now use gRPC.
  # Contracts live in api/proto/. This spec covers only the public REST surface.
components:
  securitySchemes:
    firebaseToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Firebase ID token. Validated by the Traefik forward-auth middleware
        (auth-service `/verify`), which injects `X-User-Sub` downstream.
  parameters:
...
    YearQuery:
      name: year
      in: query
      required: true
      description: Tax year, matched against the invoice date.
      schema:
        type: integer
  responses:
    Unauthorized:
      description: Missing or invalid Firebase ID token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    InvoiceCategory:
      type: string
      description: Tax-deductible expense category (German tax terms).
      enum:
        - KONTOFUEHRUNGSGEBUEHREN
        - WEGE_ZUR_ARBEIT
        - HOMEOFFICE_UND_ARBEITSZIMMER
        - INTERNET_UND_TELEFON
        - ARBEITSMITTEL
        - BERUFSVERBÄNDE_UND_GEWERKSCHAFTEN
        - STEUERBERATUNGSKOSTEN
        - REISEKOSTEN
        - BEWERBUNGEN
        - FORTBILDUNGEN
        - UMZUG
        - BEWIRTUNG
        - DOPPELTER_HAUSHALT
        - AUSSERGEWOEHNLICHE_FAHRZEUGKOSTEN
        - SONSTIGE_AUSGABEN
...
    ExportCategoryTotal:
      type: object
      required: [category, invoiceCount, total]
      properties:
        category:
          type: string
          nullable: true
          description: Null groups the invoices the AI could not categorize.
          allOf:
            - $ref: "#/components/schemas/InvoiceCategory"
        invoiceCount:
          type: integer
        total:
          type: number
    ExportSummaryResponse:
      type: object
      required:
        [year, invoiceCount, totalExpenses, lumpSumAllowance, deductibleAboveLumpSum, categories]
      properties:
        year:
          type: integer
        invoiceCount:
          type: integer
        totalExpenses:
          type: number
          description: Sum of every accepted invoice dated in this year.
        lumpSumAllowance:
          type: number
          description: >-
            The Arbeitnehmer-Pauschbetrag granted for this year without any
            receipts (§9a EStG).
        deductibleAboveLumpSum:
          type: number
          description: >-
            `max(0, totalExpenses - lumpSumAllowance)` — what the collected
            receipts are worth beyond the automatic allowance.
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ExportCategoryTotal"
...
    Error:
      type: object
      properties:
        error:
          type: string
        documentId:
          type: integer
          format: int64
          description: Present on duplicate-document (409) responses.